refactor: naming consistency
